Rename backlog.backlog to backlog.api
