+++ src/backlog/__init__.py
@@ -1,5 +1,5 @@
 """This file defines the API for backlog."""
 
-from backlog.backlog import Backlog  # noqa: F401
+from backlog.api import Backlog  # noqa: F401
 
 __version__ = '1.0.0'
